Fix an issue where focus to the previous window was not returned
Electron seems still to have an issue where focus is not properly
returned to the previous window when the main window of an application
is hidden [1].

This change introduces the workaround to this problem where the app is
first hidden, and only afterwards the main window. This hides all
windows, returns focus, and then hides the main window. The latter step
could probably be skipped, but since all checks in the code reference
`mainWindow`, it's probably better to leave that as is.

[1] https://github.com/electron/electron/issues/10023
[2] https://stackoverflow.com/a/50647468